[Style] rename variable, class
