+++ app/src/main/java/com/and04/naturealbum/ui/maps/PhotoKey.kt
@@ -1,16 +1,15 @@
-package com.and04.naturealbum.data
+package com.and04.naturealbum.ui.maps
 
-import com.and04.naturealbum.data.room.PhotoDetail
 import com.naver.maps.geometry.LatLng
 import com.naver.maps.map.clustering.ClusteringKey
 
-class ItemKey(val id: Int, private val position: LatLng) : ClusteringKey {
+class PhotoKey(val id: Int, private val position: LatLng) : ClusteringKey {
 
     override fun getPosition(): LatLng = position
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
-        if (other !is ItemKey) return false
+        if (other !is PhotoKey) return false
         return id == other.id
     }
